move to bot module
